Remove `From` impl for `Expr` to avoid API-breaking conflicts.
The `From` impl previously provided a conversion from `T` to `Expr<T>`
by making the `T` value a terminal. However, some use-cases might call
for a different conversion: for example, `T` may be another expression
AST type, and the crate's user may wish to convert the boolean
operators within it to `Expr` operators.

I think it's probably best to continue to allow this flexibility, at the
slight cost of reduced convenience. A crate that uses this one can
always implement the `From` trait as we previously had done.

Fixes #16.